skills/analysis/vlm-scene-analysis/requirements.txt
# VLM Scene Analysis
llama-cpp-python>=0.3.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
